Move ButtonsComp component from App.js to ButtonsComp.js in /components
